requirements.txt
fastapi==0.75.0
gunicorn==20.0.4
pydantic<2.0.0
requests==2.25.1
starlette==0.17.1
urllib3==1.26.4
python-dotenv==0.17.0
Hypercorn==0.11.2
pytest==7.1.1